+++ src/main/java/com/longbridge/respbodydto/ProductRespDTO.java
@@ -15,7 +15,7 @@ public class ProductRespDTO {
     public Long id;
     public String name;
     public double amount;
-    public int percentageDiscount;
+    public double percentageDiscount;
     public double slashedPrice;
     public String designerId;
     public String designerStatus;
